+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/generators/AssistedFactoryFirGenerator.kt
@@ -9,7 +9,7 @@ import dev.zacsweers.metro.compiler.fir.Keys
 import dev.zacsweers.metro.compiler.fir.MetroFirValueParameter
 import dev.zacsweers.metro.compiler.fir.classIds
 import dev.zacsweers.metro.compiler.fir.copyParameters
-import dev.zacsweers.metro.compiler.fir.findInjectLikeConstructors
+import dev.zacsweers.metro.compiler.fir.findAssistedInjectConstructors
 import dev.zacsweers.metro.compiler.fir.generateMemberFunction
 import dev.zacsweers.metro.compiler.fir.isAnnotatedWithAny
 import dev.zacsweers.metro.compiler.fir.predicates
@@ -50,7 +50,7 @@ internal class AssistedFactoryFirGenerator(session: FirSession, compatContext: C
 
   override fun FirDeclarationPredicateRegistrar.registerPredicates() {
     register(
-      session.predicates.allInjectAnnotationsPredicate,
+      session.predicates.assistedAnnotationPredicate,
       session.predicates.assistedAnnotationPredicate,
     )
   }
@@ -127,7 +127,7 @@ internal class AssistedFactoryFirGenerator(session: FirSession, compatContext: C
     classSymbol: FirClassSymbol<*>,
     context: NestedClassGenerationContext,
   ): Set<Name> {
-    val constructor = classSymbol.findInjectLikeConstructors(session).firstOrNull()
+    val constructor = classSymbol.findAssistedInjectConstructors(session).firstOrNull()
 
     if (constructor != null) {
       // Check if there is already a nested factory. If there is, do nothing.